requirements.txt
asgiref==3.5.2
autopep8==2.0.0
Django==4.1.3
djangorestframework==3.14.0
Faker==4.14.2
Pillow==9.3.0
psycopg2==2.9.5
pycodestyle==2.9.1
python-dateutil==2.8.1
pytz==2020.4
six==1.15.0
sqlparse==0.4.3
text-unidecode==1.3
tomli==2.0.1
validate-docbr==1.8.2
